Gets oscillation cycle numbers for a midline data set

Extracts the phase for a specific point along the body (defined by pointval) and uses get_cycle_numbers() to get the cycle number. Joins the cycle number data with the original data set so that cycle number is defined for each point along the body

Usage

get_body_cycle_numbers_df(
  .data,
  ph,
  pointval,
  .frame = frame,
  .point = point,
  .out = NULL,
  overwrite = TRUE,
  ...
)

Arguments

.data

Data frame containing the midlines

ph

Phase variable

pointval

Specific point on the body to use to define the phase and and cycle number

.frame, .point

Columns identifying frames and points (defaults are frame and point)

.out

Name of the output column. Needs to have one element specifying the name for the cycle variable, or as a named list with an element named cycle). Default is 'cycle'

overwrite

TRUE or FALSE to overwrite existing columns, if present.

...

Extra parameters to supply to get_cycle_numbers()

Value

A data frame containing a new column with the cycle numbers named 'cycle' or as specified in .out.